RedrawManager Update (Thanks Sidiousious ♥) Adjustments: - Put ConnectionKindChanged in Sundesmo. - Added custom scoping to redraw manager - RedrawManager performs correct redraws - Replaced the event calls from the handlers by passing the manager into them instead. - Allowed executions to hold more than one scope, and also holding their own scope inside of a RunOnPendingRedrawSlim call. - Let OwnedObjectHandlers run async with PlayerHandler, deferring their redraw results to a queue if finished before the player.

+++ Sundouleia/UI/Debug/DebugStorageUI.cs
@@ -121,8 +121,8 @@ public partial class DebugStorageUI : WindowMediatorSubscriberBase
                 ImGui.Text(incRequest.TimeToRespond.ToTimeSpanStr());
                 ImGui.TableNextColumn();
                 ImGui.Text(incRequest.Message ?? "N/A");
-                DrawIconBoolColumn(incRequest.SentFromWorld((ushort)PlayerData.CurrentWorldId));
-                DrawIconBoolColumn(incRequest.SentFromCurrentArea((ushort)PlayerData.CurrentWorldId, PlayerContent.TerritoryID));
+                DrawIconBoolColumn(incRequest.SentFromWorld(PlayerData.CurrentWorldId));
+                DrawIconBoolColumn(incRequest.SentFromCurrentArea(PlayerData.CurrentWorldId, PlayerContent.TerritoryID));
             }
         }
